Use viable default for visibility, handle all 36 bits.

+++ easycoder/CardLoader.java
@@ -54,6 +54,6 @@ public class CardLoader extends BRTLoader {
 	void endSeg() {
 	}
 
-	void beginSeg(String rev, String prg, String seg, int vis) {
+	void beginSeg(String rev, String prg, String seg, long vis) {
 	}
 }
