+++ hw2000/PeriphLoader.java
@@ -17,6 +17,6 @@ public class PeriphLoader extends BRTLoader {
 	void endSeg() {
 	}
 
-	void beginSeg(String rev, String prg, String seg, int vis) {
+	void beginSeg(String rev, String prg, String seg, long vis) {
 	}
 }
